+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Features:
 
 - Adds a circuit breaker to all requests (#72)
+- Adds resource url to the error messages
 
 ### 3.2.0 (2017-11-10)